add local sam handler
