+++ template.yaml
@@ -18,7 +18,7 @@ Resources:
       BuildMethod: go1.x
     Properties:
       CodeUri: lambda-handler/hello-world-handler
-      Handler: hello-world-handler
+      Handler: local.go
       Runtime: provided.al2
       Architectures:
         - x86_64
@@ -87,7 +87,7 @@ Resources:
       BuildMethod: go1.x
     Properties:
       CodeUri: lambda-handler/save-document-handler
-      Handler: save-document-handler
+      Handler: local.go
       Runtime: provided.al2
       Architectures:
         - x86_64
@@ -231,3 +231,28 @@ Resources:
           Properties:
             Path: /shared/{documentId}
             Method: GET
+
+  saveVectorEmbeddingsFunction:
+    Type: AWS::Serverless::Function
+    Properties:
+      CodeUri: lambda-handler/save-vector-embeddings-handler
+      Handler: local.py
+      Runtime: python3.9
+      Architectures:
+        - x86_64
+      MemorySize: 1024
+      Timeout: 100
+      Policies:
+        - DynamoDBCrudPolicy:
+            TableName: tnn-documents
+      Environment:
+        Variables:
+          USE_LOCAL_DYNAMODB: 1
+          BUCKET_NAME: faissIndexBucket
+      Events:
+        Api:
+          Type: Api
+          Properties:
+            Path: /vector-embeddings
+            Method: POST
+
